cli/requirements.txt
cmd-fwk==1.2.1
requests==2.27.1
tqdm==4.63.0